+++ src/index.ts
@@ -5,3 +5,4 @@ export * from './acs/middleware';
 export * from './acs/interfaces';
 export * from './acs/cache';
 export * from './utils';
+export * from './acs/errors';